refactor: Update ActionBtns component to include reset functionality
This commit updates the ActionBtns component to include a reset functionality. The handleResetNewTodo function is now implemented to reset the newTodo state and setChecked state to their initial values. This allows users to easily reset the form fields in the new todo component. The changes also include updating the ActionBtnsProps interface to include the initialNewTodo and setChecked props. Overall, this refactor improves the usability of the new todo feature.