refactor: Update TodoSliceStateValues interface and add TagValues interface
This commit updates the TodoSliceStateValues interface in types.d.ts to include a new property called "loading" of type boolean and another property called "error" of type boolean. It also adds a new interface called TagValues with properties "id" of type string and "name" of type string. These changes enhance the type definitions for the TodoSliceStateValues interface and provide better support for managing loading and error states in the application.