refactor: Include today's todos in TodoSliceStateValues
This commit adds a new property called "todayTodos" to the TodoSliceStateValues interface in types.d.ts. The "todayTodos" property is an array of Todo objects and will store the todos that are due today. This change enhances the state management of todos and allows for better organization and filtering based on due dates.