```text refactor: Add email export functionality to ExportBtns component
This commit adds the functionality to export tasks via email in the ExportBtns component. It includes the necessary code to send a POST request to the server with the user ID and date information. Upon successful response, a success toast notification is displayed. This enhancement provides users with an additional option to receive their exported tasks.